feat(billing): Locufy Flex com mensalidade e cobranca pelo uso de IA
Substitui os planos Starter/Growth/Professional por um plano unico (Flex,
R$ 69,90/mes) com cobranca pos-paga do consumo medido por uso: tarifas
versionadas, reserva de limite por conta, fechamento de ciclo e extrato.

- Medicao por request (ContextoIAMiddleware) e por job (por_conta), com
  autorizacao antes de chamar LLM/TTS/STT/musica e finalizacao pelas
  unidades reais
- Tabelas novas (tarifas_ia, contas_consumo, usos_ia, faturas_consumo,
  eventos_cobranca) criadas pela migracao aditiva 20260925_01 no startup
- Combinacoes texto + voz escolhidas por programa/locutor, com exemplos
  de audio no painel
- Cache de TTS e de classificacoes para reduzir custo
- Checkout e webhook do Stripe usam STRIPE_PRICE_ID_FLEX; SDK stripe 15.4
- Landing e /billing passam a oferecer apenas o Flex

IA_MEDICAO_HABILITADA e IA_ORCAMENTO_BLOQUEAR ficam desligados por padrao:
ligados sem tarifa publicada, toda geracao responde 503 e contas em trial
recebem 402. Ligar por env depois de publicar as tarifas contratadas.

+++ backend/app/admin_sistema/router.py
@@ -29,7 +29,7 @@ class OverviewResponse(BaseModel):
     novas_empresas_30_dias: int
     total_usuarios_ativos: int
     mensagens_30_dias: int
-    mrr_planos: int
+    mrr_planos: float
     mrr_agentes_extras: int
